Reorganize Sidebar component files.
