+++ new-log-viewer/src/components/CentralContainer/Sidebar/PanelTabs/CustomListItem.tsx
