+++ new-log-viewer/src/components/CentralContainer/Sidebar/PanelTabs/CustomTabPanel.tsx
