+++ new-log-viewer/src/components/CentralContainer/Sidebar/PanelTabs/TooltipTab.css
