+++ new-log-viewer/src/components/CentralContainer/Sidebar/PanelTabs/index.css
