+++ new-log-viewer/src/components/CentralContainer/Sidebar/ResizeHandle.tsx
